Implemented RPG::Table, Map File Support
Fixed an issue with RPG::Table objects causing a crash by finding an obscure forum post and inputting their implementation of the Table object into the code itself. Also created a map mode which I devolved for map files, but is a no nonsense throw everything into a file. Implemented another issue of rvdata2 files just containing a single object, which was the case with map files